Вынес генерацию тестовых данных в mock из main.js
